refactor chunk reading
Hound should now be able to cope with wav files with exotic chunks or chunks in
weird order, as long as the "fmt" chunk precedes the "data" chunk. Any chunks
following the data chunk will be ignored. (Unknown chunks will be skipped
anyway.) I am unable to generate a wav file with extra chunks before the data
chunk, though Audition does save files with embedded xml nonsense after the
data chunk. So, this is not tested, but at least it does not break anything.