signal error when not writing enough samples
The half-open intervals used for ranges were actually a footgun to me this
time, but in this case the error could have easily been prevented by Hound,
if it verifies upon finalization that the number of samples written is a
multiple of the number of channels. So I added the check.