test that a wide write returns an error
It is difficult, at the moment, to test whether the right kind of error is
returned, because `hound::Error` cannot implement `Eq` because `io::Error` does
not implement it. This makes it impossible to do an `assert_eq!` on the error,
and the only way to verify that it has the right kind is to do a match (or
if-let maybe), which is tedious and which would blow up the test function.