drop support for Rust 1.0.0 through 1.3.0
Either:

 * Get warnings about deprecated items on Rust >= 1.4.0.
 * Don't compile on stable Rust < 1.4.0.

On the one hand, why drop support for older Rust versions unnecessarily?
On the other hand, Hound 1.1.0 still works fine with older versions of
Rust, and I don't expect a new release of the library to be necessary
soon. If you want to use a bleeding-edge version of Hound, you have to
use the latest Rust. Seems reasonable.