add WAVEFORMATEXTENSIBLE IEEE float test sample
This ensures that all code paths are tested again. This file has been
generated with the CopyAudio utility with the following flags:

    $ CopyAudio -t WAVE -F WAVE -D float32 -S "FL" src.wav dest.wav

Setting the speaker assignment with the -S flag is required to make the
utility produce a WAVEFORMATEXTENSIBLE file; otherwise it produces a
regular WAVEFORMATEX file.

CopyAudio is part of the AFsp package [1]. I had to build it from
source, available at [2]. There are many real-world test samples
available at that site too [3], including a WAVEFORMATEXTENSIBLE IEEE
float sample. However, I did not want to include that one because it is
larger than it needs to be (a few hundred kilobytes, vs. a few hundred
bytes now), and because the site contains no licensing information.

[1]: http://www-mmsp.ece.mcgill.ca/Documents/Software/Packages/AFsp/AFsp.html
[2]: http://www-mmsp.ece.mcgill.ca/Documents/Downloads/AFsp/
[3]: http://www-mmsp.ece.mcgill.ca/Documents/AudioFormats/WAVE/Samples.html