Do not buffer writes automatically
This should be the responsibility of the caller, and the convenience
constructor for writer does wrap the writer in a buffered writer
already. As the buffered writer was an implementation detail, this is
not a breaking change, however performance might degrade in case you
were relying on this automatic buffering.