Expose unsafe write_sample_unchecked()
If you know how many samples you are going to write in advance anyway,
then you probably do that in some kind of loop, and you never write more
samples. It would be nice to be able to get rid of the bounds check
there. Unfortunately the compiler does not seem to be able to elide it.
So expose an unsafe function then.