Avoid multiplication overflow when reading header
This is not a problem in practice: in release mode the multiplication
would wrap, and the check would fail. It would be possible to craft a
malicious header for which the check would not fail, but this would not
cause other issues. However, this is still annoying in debug mode, so
let's fix it.

This issue was found by libfuzzer using cargo-fuzz.