Upgrade cpal to 0.2.12
This is still an old version, but it is the last version that does not
depend on the futures library. We have to use this older version to
still support Rust 1.4.

The newer cpal version is not as easy to use, but in return we know that
playback makes sense, and that the audio device matches the wav spec.

Hopefully this upgrade will also fix the build on OS X, which is is what
triggered the upgrade in the first place.