Put release dates in changelog
People are busy, they are not going to clone the repository and check
the dates on the tags, or even browse through releases on GitHub.
Putting this directly in the changelog allows developers to tell at a
glance whether they have been using outdated dependencies for a long
time or not, is actually something that I occasionally want to know.
So let's put it in there.