Work around breaking change in Cargo / libc crate
A newer version of the libc crate cannot be unpacked by older versions
of Cargo any more; the Cargo bundled with Rust 1.4.0 complains about an
invalid tar archive.

Version 0.2.24 of the libc crate still worked fine, but version 0.2.29
cannot be unpacked. I suspect that it was uploaded by a newer version of
Cargo which produces different archives. I should watch out myself for
publishing with a Cargo that is too new.

This issue was caught by a periodic build on Travis CI ("cron job"). The
exact same commit compiled fine about a month ago, but when it was built
again, the build failed. The Rust ecosystem is really not stable at all.

The solution is to make builds reproducible by putting the lockfile
under source control. This way we don't test compatibility with other
crates, but there are no (non-development) dependencies anyway.

See also https://github.com/rust-lang/cargo/issues/4143.