Add support for 40-byte WAVEFORMATEX fmt chunks
A file encountered in the wild sports a 40-byte fmt chunk, but the
format tag is the PCM format tag, not WAVEFORMATEXTENSIBLE. The file
furthermore has 24 bits per sample, something which is invalid according
to the standard. However, there is nothing that prevents that bit depth
from being loaded in princible, so let's support it for the sake of
compatibility.

I will try to trim down the file so it can be included as a test case in
the repository in a follow-up commit.