Add test sample for nonstandard file
This is a truncated version of a file found in the wild (pointed to by a
bug report). The actual audio data has been removed, and some surgery
has been applied to the headers with a hex editor, but the problematic
parts are still there. This test does not pass with Hound 3.1.0.