Add WavReader::seek_sample method
This adds a method to WavReader that allows the user to seek to a given
sample position within the file.

This method requires that the inner reader `R` implements `io::Seek`.

The given sample position represents the number of samples from the
beginning of the audio data. Thus the given sample position should not
exceed the length of the file in samples (returned by `len()`). The
behaviour when seeking beyond `len()` depends on the reader's `Seek`
implementation.

A test for the new method is included at the bottom of the module.

This closes #18.