Rename seek_sample to seek
Also update the docs, and rename the argument to "time". I think this is
slightly clearer, because "time" inherently does not depend on the
number of channels, whereas for "sample", it is somewhat unclear if that
refers to the value of a single channel, or the values of all channels.
It is probably best to avoid the term "sample" if possible.