Add test for second nonstandard test sample
This file is the first few bytes of a nonstandard file encountered in
the wild (apparently part of Quake in pak0.pak/sound/doors/baseuse.wav).
It has been truncated to 8 bytes of data chunk, and the audio data has
been zeroed out. I did not patch the header yet, so the length does not
match currently.

Hound cannot currently read this file, let's see if we can make sense
of it and add support.