Fix bug in reading WAVEFORMATEX cbSize member
In the specific case of WAVE_FORMAT_PCM, the WAVEFORMATEX member cbSize
should be ignored; it need not be zero. Nonzero values do occur in
practice. This fixes the bug and adds a test case. I now believe
the test file is actually valid, it is not a nonstandard file as I
previously thought.