Start working on append functionality
This is a very rough first version. The challenge lies in identifying
the header which is used in the file, and then making sure to write the
same header, and not corrupting it. In some cases this might not be
possible, when the header is one that Hound can read, but not write.