Expose read_wave_header(), move WavSpecEx
This allows those pieces to be used by WavWriter::append(), which needs
to perform a read first, and it also needs some more low-level
information about what kind of WAVE header was present than what the
writer normally provides.

This does not compile; I also need to move the part that reads up to the
data chunk.