Properly handle headers in case of append
Handle headers by recording the position of the data chunk length field,
so we can overwrite it at the proper location. This works even for
headers that Hound does not write itself.

As a drive-by fix, it turns out Hound did not write the WAVEFORMATEX
struct, but rather the older PCMWAVEFORMAT struct. This is not an issue
in itself, but the naming was wrong. Also, MSDN says the PCMWAVEFORMAT
struct has been "superseded" by WAVEFORMATEX, but in practice any
reading library that supports WAVEFORMATEX would likely support both.
So let's keep using the older struct.