Add convenience constructor WavWriter::append()
And rename the previous WavWriter::append() to WavWriter::new_append().

It turns out that an io::BufWriter<W> where W: io::Read + io::Write,
does not implement io::Read, so simply wrapping a file in a BufWriter
and passing it into WavWriter::new_append() does not work. We have to
use a separate BufReader and BufWriter.