Do not build append example on Rust 1.4.0
It uses features that Rust 1.4.0 did not have. They can be worked
around, but in the case of the example there is no good reason to;
we don't want to break existing code, but this is a new example,
we are not causing a compile error in code that compiled previously.