Introduce ChunksWriter to support ad-hoc chunks
ChunksWriter basically inherits all the code that used to be in
WavWriter. The WavWriter is now a facade for ChunksWriter,
preserving the existing API and preventing the ChunksWriter to
reach invalid states.

An EmbeddedWriter struct is added allowing to write an arbitrary chunk.
It can be obtained by calling start_chunk on the ChunksWriter. The data
chunk is opened by calling start_data_chunk.

The chunk header maintenance routines for both the datachunk managed by
ChunksWriter itself and EmbeddedWriters are shared in an the
ChunkWritingState struct, mimicking the Reader case.