Add S24_LE wav files and tests
Add two test files, one using pcmwaveformat with bitsPerSample == 24,
and one using waveformatextensible with validBitsPerSample == 24. Also
add tests that verify these two files are properly read.

According to the standard, the first sort of file should not be
permitted, but it is created by arecord, so it is a good type of file to
try and support.

(cherry picked from commit 99c120babad08179cd47000f62d82f37f4e71cee)