Pick Rust 1.56 as the MSRV
I was always erring on the side of being compatible, but times have
changed. People who are stuck with older Rust toolchains can use an
older Hound version just fine. For Hound itself, I think it is now time
to move on to something more recent, and adopt some of the new stuff.

If I'm going to make big breaking changes and people need to change
their programs anyway, then it's fine to also demand a newer Rust
compiler. The people who are unable to use it can just not use the new
version. In the wider ecosystem, 1.56 is pretty old by now, so even if
Hound is pretty low in the stack, I think it should be okay. It will
take me some time to write this, it will take the ecosystem some time
to adopt it -- if they will adopt it at all.