Add reminder to expose sample format later
Ambisonic is not linear PCM, if you can just load it, don't check for
this type, and then try to play it as if it's linear PCM, you might hear
garbage, or worse, damage your speakers, so a better API is needed for
this. But that is an invasive thing that I don't want to change right
now.